Changing interval option only
Changing interval option to be 0. Nothing else was changed in the tests. This is important, since the tests assume typing two consecutive numbers to be separate transactions. With the default interval (1000), they should be considered a single transaction. Either change the test with a delay > 1000ms (or whatever the interval is), or what I did, change the interval to be 0.