Refactoring and updating _previousContent properly
An important refactoring and update. _previousContent is now updated using getHTML from inside setHTML. It makes sure the current and the previous content are using the same method to get the content html. Also, setting _previousContent inside setHTML, ensures it gets updated when undoing and redoing automatically (since it is used by restoreFromHistory).