Removing a no longer needed measure
Force pushing a history on focus is no longer needed with the new undo manager. When scribe is instantiated, its initial content is saved in _previousContent. Also, whenever setContent is called, the new content is also saved in _previousContent. This means the next real change transaction will be aware of the previous content whether it was from an empty scribe, or one that was set using setContent.