Improving UndoManager space and time efficiency
Before this update, each history item stored both the previous and the current content. This means there are redundancies between every pair of history items. This update removed the need to store the previous content by only storing a reference to the previous item. This way, only current content is saved in each history item. Now we are storing a reference to the last history item to improve access time and content retrieval/update. Memory space cannot be saved more than this, unless a diff algorithm is used to only store the differences (which can be inefficient time-wise, and more complex with undo/redo)