Build: switch from Gulp to Grunt
Gulp ended up being more complex that the value it was adding because most of the compilers we are using are not stream orientated.

We need to switch the build process to be something that will move us to the latest Node releases as we want to start switching to using ES2015 features as they become available in the browser.

I have also removed the SauceLabs add-on from the Travis configuration as the Guardian no longer has an organisational account and the browser-based testing needs a rethink anyway.